stash conflict fix
